Initial loading works - but overflow crash
The killer thing to get around was that the handler needed ctx passed to it (I didn't see this in the demos I was following for some reason).

Bug at the moment as I'm exceeding the buffers - found that there is a built in limit of 124bytes (thanks Katherine for getting back to me so quick!) - so I need to find a way to get around the limit (needs more investigation).